add team on CFQ guest
